gkzRestApi/requirements.txt
django==2.1.7
djangorestframework
pymysql
mysqlclient